fuzzers/qemu_launcher: Use abort instead of exit
Previously, if you didn't have a corpus and uncommented the line about
redirecting stdout to `/dev/null`, you would see:
```
Failed to load initial corpus at ["./corpus"]
thread 'main' panicked at 'Fuzzer-respawner: Storing state in crashed
fuzzer instance did not work, no point to spawn the next client! This
can happen if the child calls `exit()`, in that case make sure it uses
`abort()`, if it got killed unrecoverable (OOM), or if there is a bug in
the fuzzer itself. (Child exited with: 0)',
```