Introduce `AsCopyIter`; towards map observers in safe Rust
Previously, `MapFeedback` required the `Observer` to implement `AsIter`.
This works for all the in-repo `MapObserver`s, which use `unsafe` to
pretend that they own the underlying map/vector/array. This lets them
loan out items in the array with the same lifetime as `&self`.

It would be nice to have a `MapObserver` built in safe Rust out of
`ValueObserver<RefCell<Vec<...>>>` or a similar type with interior
mutability.  If such a type implemented `MapObserver`, it could be used
with the map feedbacks, like `MaxMapFeedback`. However, the borrow from
a `RefCell` by necessity has a shorter lifetime than `&self`, so a type
like the above can't implement `AsIter`.

`MapFeedback` already requires that all the values are `Copy`, and in
practice, they're generally small (in fact, they're generally primitive
integers, meaning copying is free). Thus, `AsCopyIterator` creates
iterators that just copy the values out of the container, avoiding
questions about lifetimes and borrows.