`NopScheduler`, for use in generative fuzzers
This commit introduces the `NopScheduler`, which never picks a testcase.
This is useful for generative fuzzers, which don't maintain a corpus.

Previously, `StdFuzzer::fuzz_one` would panic if the scheduler
didn't return a `CorpusId`, but now it handles that case by ignoring
`Error::Empty` and returning an `Option<CorpusId>` itself.