fix(input-on-crash): Also add non objectives as inputs after crash
Previously, the crash handler in the inprocess executor only handled
adding inputs that are considered an objective. But when working with
non-crash objectives this leads to the fuzzer never adding the input and
ignoring it.

This commit fixes this behavior by introducing a new function
to the `Evaluator` trait & adjusting some:

- Added `evaluate_input_after_execute`, which does the same as
  `evaluate_input_with_observers` but without executing the input
- We reused the existing evaluation functions, because we
  want to use the evaluation logic there without executing the input
  again. `run_observers_and_save_state` now just calls the method on the
  fuzzer directly.
- The `evaluate_input` method was adjusted to also use the same logic
- We switched the type constraint on the Fuzzer generic to the
  `Evaluator` trait to be able to access the new method and thus pushed
  those requirements further where needed.

Note: We also mostly removed the Feedback generic parameters `OF` where possible, since those are not needed anymore.