Separate state trait/impl for generative fuzzers
Generative fuzzers do not need a corpus. To this end, this commit splits
out `GenState` (trait) and `StdGenState` (struct) from `State` (trait)
and `StdState` (struct). The idea is that we will support some notion of
a generative fuzzer (perhaps `GenFuzzer`) that will not rely on having a
"current corpus ID". Down the line, we may also want to move input
loading/generation/scheduling into a method on `GenState` that returns
a `State`, see #2200.