Fixes in corpus minimizer (#3403)
* Fix removing file with ctr of the form "\u{3}000"

* Ensure input is loaded, also ensure exec_time is set

* Clippy

* Revert "Fix removing file with ctr of the form "\u{3}000""

This reverts commit 4cb3507405e726538df94591ecd5c851c8778efc.

* libafl: fix lockfile ctr serialization (#3319)

All locations reading from the lockfile expect an integer as a string,
but in `InMemoryOnDiskCorpus<I>::remove_testcase` the raw bytes of the
integer are written to the file in little endian byte order.

This may cause "ParseIntError { kind: InvalidDigit }" in e.g.
`InMemoryOnDiskCorpus<I>::save_testcase` during parsing of the file's
contents.

Fix this by writing the integer to the lockfile as a string in
`InMemoryOnDiskCorpus<I>::remove_testcase` as well.

* Extract run target with timing

* Expose run_target_with_timing

* Simplify

* HasExecutions

* clippy

* linters

* fmt

---------

Co-authored-by: Niklas Gögge <n.goeggi@gmail.com>