Update libpng versions and fix MacOS failures
libpng (and zlib) was not building properly on the latest MacOS due to redifinition of fopen.
With the new version of libpng and zlib it builds without previous problems, but `make` fails with a new problem with feenableexcept, but in tests only. But since that in libpng's tests, I just made autotools "disable" the option.