feat: add just docs building for libafl_cc, libafl_derive, libafl_nyx
Some intradoc links had to be fixed in libafl_nyx.
Building the docs for this also requires cpio.
There are some issues with the docs of libafl_nyx i.e.
the README.md file does not get added to the docs.
After this commit, I am going to build every single doc
again and then go through everything built to check that
all links work, there are no typos and that all relevant
info gets added to the docs.